app/requirements.txt
# Django
Django==4.2
psycopg2-binary==2.9.9
environs==11.0.0

# Telegram Bot
aiogram==3.13.1

# AI Services
openai==1.54.3
google-generativeai==0.8.3

# Document Generation
python-docx==1.1.2

# Redis
redis==5.2.0

# Web Server
gunicorn==23.0.0

# Utilities
python-dateutil==2.9.0
